+++ db/seeds_demo_v1.rb
@@ -19,14 +19,17 @@ puts <<~MSG
 
 [demo] Готово:
   Организация: #{result.organization.name} (#{result.organization.slug})
+  Организация Alt: #{result.organization_alt.name} (#{result.organization_alt.slug})
   Точка A: #{result.tenant_a.name} (#{result.tenant_a.slug}) id=#{result.tenant_a.id}
   Точка B: #{result.tenant_b.name} (#{result.tenant_b.slug}) id=#{result.tenant_b.id}
+  Точка C: #{result.tenant_c.name} (#{result.tenant_c.slug}) id=#{result.tenant_c.id}
   Заготовочный цех: #{result.tenant_kitchen.name} (#{result.tenant_kitchen.slug}) id=#{result.tenant_kitchen.id}
   Активных продуктов: #{result.products_count}
-  PTS: A=#{result.pts_a_count}, B=#{result.pts_b_count}
+  PTS: A=#{result.pts_a_count}, B=#{result.pts_b_count}, C=#{result.pts_c_count}
   Demo-пользователей: #{result.users_count} (в т.ч. pk-manager / pk-worker → /prep_kitchen)
   Пароль: #{Demo::EnvironmentSetup::DEMO_PASSWORD}
   Shop A: #{demo_shop_public_url(result.tenant_a)}
   Shop B: #{demo_shop_public_url(result.tenant_b)}
+  Shop C: #{demo_shop_public_url(result.tenant_c)}
 
 MSG